requirements.txt
Flask==3.0.3
gunicorn==21.2.0
SQLAlchemy==2.0.30
psycopg[binary]==3.1.20
requests==2.32.3
python-dotenv==1.0.1
